+++ NEWS.md
@@ -1,10 +1,12 @@
-# Version 0.1-4, 2020-08-17
+# Version 0.1-4, 2020-08-23
 
 * `.edma_fit_np` is more preformant (original implementation
   retained as `.edma_fit_np_old` for comparison);
   it also exposes mean/variance for distances for future uses.
 * `edma_simulate_data` is also much faster now.
 * Nonparametric EDMA fid and FDM analysis can use multiple cores.
+* Added `write.xyz` function to write EDMA data into xyz format.
+* Added 4 data sets for growth difference analysis.
 
 # Version 0.1-3, 2020-07-05